Add a NotificationRestService and transforms of notifications
Implement a new NotificationRestService that is a HubRestService. 
Implement a NotificationDataService class that will return a list of
objects containing data from the hub.  These objects are intended to be
used by the integrations without having to make additional rest calls.
Implement some transform objects to transform NotificationItems into the
objects the NotificationDataService will provide to the CI.