Not all items can have a limit and offset by default workaround it
Add an integer value -999 to denote that the integer value of limit and
offset can be excluded from the HubRequest.  This will be removed with a
new type of hub request that will be implemented in the future.
Therefor the TODO comment was added so that it is not forgotten.