Make Rest Connection abstract and add a CredentialsRestConnection
Make the Rest Connection class abstract in order to remove the need for
a HubServerConfig object to be populated correctly to create the rest
connection.  Create a CredentialsRestConnection that will take the
HubServerConfig object in the constructor and initialize the
RestConnection object. If the username and password are empty then do
not call the setCookies method in the constructor.