IHCIC-81 Change the file paths to be strings using '/' character.
When looking for resources need to use a forward / character.  Change
the code to use strings for the relative paths of the resources in order
to locate them using getResourceAsStream.  Also try using the
getClass().getResourceAsStream() method rather than the classloader.  We
will try both approaches in order to load the risk report resources.