Fix the transformer to create the policy violation event to add
If components violating policy are cleared or overridden the policy
violation event is never generated and it is skipped.  This causes a
problem where you don't have matching policy violation event and policy
override or clear events that cancel each other out in a time range.
This can result in reporting incorrect information.  To fix this remove
the checks