Refactor the code to clean up some outstanding items.
Clean up the DetailFactory to use a functional programming to
encapsulate the detail construction into one common method.  The only
difference is the method used to create the detail list.

Fix the TODO on the NotifciationResults to simply use a generic type to
define the return of the getResults list since the two potential return
types have nothing in common as far as an API goes.