Consider BDS_JAVA_HOME for Signature Scanner Execution
Executing Signature Scanner from Blackduck Hub currently fails on
Alpine Linux environments as the Oracle JRE packaged with Signature
Scanner requires glibc, which is not present on Alpine Linux. They use
musl libc instead.

Executing Signature Scanner with the systems default OpenJDK JRE is not
possible as hub-common uses the Oracle JDK packaged with Signature
Scanner.

Starting Signature Scanner from command line works, as it considers the
environment variable BDS_JAVA_HOME and takes the JRE located there
instead of the Oracle JRE pacakged with Signature Scanner.

Unit tests are done with PowerMock instead of a wrapper class for
java.lang.System so the constructor of ScanPathsUtility stays the same
as it is used by hub-detect and as such changing it would cause
incompatibility. We had to use a beta version of PowerMock as others
are not yet compatible with Mockito2.

Also added a test for specific output directory.

Fixes #224